requirements.txt
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1